refactor(stats): extract token-budget and index stats modules (#677)
Split src/stats.ts (944 -> 175-line facade) by responsibility:
- stats-tokens.ts: median, computeTokenBudget, formatTokenBudgetReport
- stats-index.ts: repo/author/index stats + their report formatters
Shared test fixtures (makeSkill/emptyAuditReport/makeRepoIndex) move to
stats-test-fixtures.ts; describes move to stats-index/stats-tokens test
files. Facade re-exports the full prior surface; no importer changes.
