+++ src/stats.test.ts
@@ -1,69 +1,16 @@
-import { describe, expect, it, beforeEach, afterEach } from "vitest";
+import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import {
   dirSize,
   computeStats,
   formatHumanSize,
   formatStatsReport,
-  computeRepoStats,
-  computeAuthorStats,
-  computeIndexStats,
-  formatRepoStatsReport,
-  formatAuthorStatsReport,
-  formatIndexStatsReport,
-  computeTokenBudget,
-  formatTokenBudgetReport,
-  median,
 } from "./stats";
-import type {
-  SkillInfo,
-  AuditReport,
-  StatsReport,
-  RepoIndex,
-  RepoStatsReport,
-  AuthorStatsReport,
-  IndexStatsReport,
-} from "./utils/types";
+import type { StatsReport } from "./utils/types";
 import { mkdtemp, writeFile, mkdir, rm } from "fs/promises";
 import { join } from "path";
 import { tmpdir } from "os";
-
+import { makeSkill, emptyAuditReport } from "./stats-test-fixtures";
 let tempDir: string;
-
-function makeSkill(overrides: Partial<SkillInfo> = {}): SkillInfo {
-  const path = overrides.path ?? "/tmp/test-skill";
-  return {
-    name: "test-skill",
-    version: "1.0.0",
-    description: "A test skill",
-    creator: "",
-    license: "",
-    compatibility: "",
-    allowedTools: [],
-    dirName: "test-skill",
-    path,
-    originalPath: path,
-    location: "global-claude",
-    scope: "global",
-    provider: "claude",
-    providerLabel: "Claude Code",
-    isSymlink: false,
-    symlinkTarget: null,
-    realPath: path,
-    fileCount: 3,
-    effort: undefined,
-    ...overrides,
-  };
-}
-
-function emptyAuditReport(): AuditReport {
-  return {
-    scannedAt: new Date().toISOString(),
-    totalSkills: 0,
-    duplicateGroups: [],
-    totalDuplicateInstances: 0,
-  };
-}
-
 describe("dirSize", () => {
   beforeEach(async () => {
     tempDir = await mkdtemp(join(tmpdir(), "stats-test-"));
@@ -244,825 +191,6 @@ describe("formatStatsReport", () => {
     );
     expect(output).toContain("0 skills");
     expect(output).toContain("0 B");
-  });
-});
-
-// ─── computeRepoStats ────────────────────────────────────────────────────────
-
-function makeRepoIndex(overrides: Partial<RepoIndex> = {}): RepoIndex {
-  return {
-    repoUrl: "https://github.com/test/repo",
-    owner: "test",
-    repo: "repo",
-    updatedAt: new Date().toISOString(),
-    skillCount: 0,
-    skills: [],
-    ...overrides,
-  };
-}
-
-describe("computeRepoStats", () => {
-  it("returns empty array for no indices", () => {
-    const result = computeRepoStats([]);
-    expect(result).toEqual([]);
-  });
-
-  it("computes stats for a single repo", () => {
-    const index = makeRepoIndex({
-      owner: "anthropics",
-      repo: "skills",
-      skills: [
-        {
-          name: "code-review",
-          description: "Code review best practices",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: ["read", "edit", "str_replace_editor"],
-          installUrl: "github:anthropics/skills:skills/code-review",
-          relPath: "skills/code-review",
-          verified: true,
-          tokenCount: 1200,
-        },
-        {
-          name: "openspec",
-          description: "OpenSpec workflow for spec-driven development",
-          version: "0.2.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: ["read", "write", "edit"],
-          installUrl: "github:anthropics/skills:skills/openspec",
-          relPath: "skills/openspec",
-          verified: false,
-          tokenCount: 3500,
-        },
-      ],
-    });
-
-    const result = computeRepoStats([index]);
-    expect(result).toHaveLength(1);
-    expect(result[0].owner).toBe("anthropics");
-    expect(result[0].repo).toBe("skills");
-    expect(result[0].skillCount).toBe(2);
-    expect(result[0].verifiedCount).toBe(1);
-    expect(result[0].totalTokens).toBe(4700);
-    expect(result[0].categories["coding"]).toBe(1);
-  });
-
-  it("sorts repos by skill count descending", () => {
-    const indexA = makeRepoIndex({
-      owner: "user",
-      repo: "small",
-      skills: [
-        {
-          name: "a",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user/small:a",
-          relPath: "a",
-        },
-      ],
-    });
-    const indexB = makeRepoIndex({
-      owner: "user",
-      repo: "big",
-      skills: [
-        {
-          name: "a",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user/big:a",
-          relPath: "a",
-        },
-        {
-          name: "b",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user/big:b",
-          relPath: "b",
-        },
-        {
-          name: "c",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user/big:c",
-          relPath: "c",
-        },
-      ],
-    });
-
-    const result = computeRepoStats([indexA, indexB]);
-    expect(result[0].repo).toBe("big");
-    expect(result[1].repo).toBe("small");
-  });
-
-  it("handles skills with evalSummary", () => {
-    const index = makeRepoIndex({
-      owner: "test",
-      repo: "eval-repo",
-      skills: [
-        {
-          name: "skill-a",
-          description: "A skill",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "test",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:test/eval-repo:skill-a",
-          relPath: "skill-a",
-          evalSummary: {
-            overallScore: 90,
-            grade: "A",
-            categories: [],
-            evaluatedAt: new Date().toISOString(),
-          },
-        },
-        {
-          name: "skill-b",
-          description: "B skill",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "test",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:test/eval-repo:skill-b",
-          relPath: "skill-b",
-          evalSummary: {
-            overallScore: 75,
-            grade: "C",
-            categories: [],
-            evaluatedAt: new Date().toISOString(),
-          },
-        },
-      ],
-    });
-
-    const result = computeRepoStats([index]);
-    expect(result[0].avgEvalScore).toBe(83);
-  });
-});
-
-// ─── computeAuthorStats ──────────────────────────────────────────────────────
-
-describe("computeAuthorStats", () => {
-  it("returns empty array for no indices", () => {
-    const result = computeAuthorStats([]);
-    expect(result).toEqual([]);
-  });
-
-  it("aggregates across multiple repos for same author", () => {
-    const index1 = makeRepoIndex({
-      owner: "anthropics",
-      repo: "skills",
-      skills: [
-        {
-          name: "code-review",
-          description: "Code review",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:anthropics/skills:cr",
-          relPath: "cr",
-        },
-      ],
-    });
-    const index2 = makeRepoIndex({
-      owner: "anthropics",
-      repo: "more-skills",
-      skills: [
-        {
-          name: "openspec",
-          description: "OpenSpec workflow",
-          version: "0.2.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:anthropics/more-skills:os",
-          relPath: "os",
-        },
-        {
-          name: "eval",
-          description: "Eval best practices",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:anthropics/more-skills:eval",
-          relPath: "eval",
-        },
-      ],
-    });
-
-    const result = computeAuthorStats([index1, index2]);
-    expect(result).toHaveLength(1);
-    expect(result[0].owner).toBe("anthropics");
-    expect(result[0].totalSkills).toBe(3);
-    expect(result[0].repos).toContain("anthropics/skills");
-    expect(result[0].repos).toContain("anthropics/more-skills");
-    expect(result[0].repos).toHaveLength(2);
-  });
-
-  it("sorts authors by total skills descending", () => {
-    const indexA = makeRepoIndex({
-      owner: "big-author",
-      repo: "repo-a",
-      skills: [
-        {
-          name: "a",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:big-author/repo-a:a",
-          relPath: "a",
-        },
-        {
-          name: "b",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:big-author/repo-a:b",
-          relPath: "b",
-        },
-      ],
-    });
-    const indexB = makeRepoIndex({
-      owner: "small-author",
-      repo: "repo-b",
-      skills: [
-        {
-          name: "x",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:small-author/repo-b:x",
-          relPath: "x",
-        },
-      ],
-    });
-
-    const result = computeAuthorStats([indexA, indexB]);
-    expect(result[0].owner).toBe("big-author");
-    expect(result[1].owner).toBe("small-author");
-  });
-
-  it("handles multiple authors", () => {
-    const indexA = makeRepoIndex({
-      owner: "user-a",
-      repo: "repo-a",
-      skills: [
-        {
-          name: "a",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user-a/repo-a:a",
-          relPath: "a",
-        },
-      ],
-    });
-    const indexB = makeRepoIndex({
-      owner: "user-b",
-      repo: "repo-b",
-      skills: [
-        {
-          name: "b",
-          description: "",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:user-b/repo-b:b",
-          relPath: "b",
-        },
-      ],
-    });
-
-    const result = computeAuthorStats([indexA, indexB]);
-    expect(result).toHaveLength(2);
-    const owners = result.map((r) => r.owner);
-    expect(owners).toContain("user-a");
-    expect(owners).toContain("user-b");
-  });
-});
-
-// ─── computeIndexStats ───────────────────────────────────────────────────────
-
-describe("computeIndexStats", () => {
-  it("returns empty stats for no indices", () => {
-    const result = computeIndexStats([]);
-    expect(result.totalRepos).toBe(0);
-    expect(result.totalSkills).toBe(0);
-    expect(result.totalAuthors).toBe(0);
-    expect(result.verifiedCount).toBe(0);
-    expect(result.avgTokensPerSkill).toBe(0);
-  });
-
-  it("computes aggregate stats across all indices", () => {
-    const index1 = makeRepoIndex({
-      owner: "anthropics",
-      repo: "skills",
-      skills: [
-        {
-          name: "code-review",
-          description: "Code review",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:anthropics/skills:cr",
-          relPath: "cr",
-          verified: true,
-          tokenCount: 1200,
-        },
-        {
-          name: "openspec",
-          description: "OpenSpec workflow",
-          version: "0.2.0",
-          license: "MIT",
-          creator: "anthropics",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:anthropics/skills:os",
-          relPath: "os",
-          tokenCount: 3500,
-        },
-      ],
-    });
-    const index2 = makeRepoIndex({
-      owner: "luongnv89",
-      repo: "asm",
-      skills: [
-        {
-          name: "hello-world",
-          description: "Hello world skill",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "luongnv89",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:luongnv89/asm:h",
-          relPath: "h",
-          tokenCount: 500,
-        },
-      ],
-    });
-
-    const result = computeIndexStats([index1, index2]);
-    expect(result.totalRepos).toBe(2);
-    expect(result.totalSkills).toBe(3);
-    expect(result.totalAuthors).toBe(2);
-    expect(result.verifiedCount).toBe(1);
-    expect(result.totalTokens).toBe(5200);
-    expect(result.avgTokensPerSkill).toBe(1733);
-  });
-
-  it("computes category distribution", () => {
-    const index = makeRepoIndex({
-      owner: "test",
-      repo: "repo",
-      skills: [
-        {
-          name: "code-review",
-          description: "Code review best practices",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "test",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:test/repo:cr",
-          relPath: "cr",
-        },
-        {
-          name: "docker-setup",
-          description: "Docker containerization",
-          version: "1.0.0",
-          license: "MIT",
-          creator: "test",
-          compatibility: "claude",
-          allowedTools: [],
-          installUrl: "github:test/repo:ds",
-          relPath: "ds",
-        },
-      ],
-    });
-
-    const result = computeIndexStats([index]);
-    expect(result.categoryDistribution["coding"]).toBe(1);
-    expect(result.categoryDistribution["devops"]).toBe(1);
-  });
-});
-
-// ─── formatRepoStatsReport ───────────────────────────────────────────────────
-
-describe("formatRepoStatsReport", () => {
-  beforeEach(() => {
-    (globalThis as any).__CLI_NO_COLOR = true;
-  });
-  afterEach(() => {
-    delete (globalThis as any).__CLI_NO_COLOR;
-  });
-
-  function makeReport(
-    overrides: Partial<RepoStatsReport> = {},
-  ): RepoStatsReport {
-    return {
-      owner: "anthropics",
-      repo: "skills",
-      repoUrl: "https://github.com/anthropics/skills",
-      skillCount: 5,
-      categories: { coding: 2, testing: 1, security: 1, devops: 1 },
-      verifiedCount: 3,
-      totalTokens: 10000,
-      avgEvalScore: 85,
-      ...overrides,
-    };
-  }
-
-  it("includes repo name and overview", () => {
-    const output = formatRepoStatsReport(makeReport());
-    expect(output).toContain("Repo: anthropics/skills");
-    expect(output).toContain("Skills:");
-    expect(output).toContain("5");
-    expect(output).toContain("Verified:");
-    expect(output).toContain("3");
-    expect(output).toContain("Avg Eval:");
-    expect(output).toContain("85");
-  });
-
-  it("shows category distribution with bars", () => {
-    const output = formatRepoStatsReport(makeReport());
-    expect(output).toContain("Categories");
-    expect(output).toContain("coding");
-    expect(output).toContain("testing");
-    expect(output).toContain("#");
-  });
-
-  it("handles zero skills", () => {
-    const output = formatRepoStatsReport(
-      makeReport({ skillCount: 0, categories: {}, verifiedCount: 0 }),
-    );
-    expect(output).toContain("Repo: anthropics/skills");
-    expect(output).toContain("0");
-  });
-});
-
-// ─── formatAuthorStatsReport ─────────────────────────────────────────────────
-
-describe("formatAuthorStatsReport", () => {
-  beforeEach(() => {
-    (globalThis as any).__CLI_NO_COLOR = true;
-  });
-  afterEach(() => {
-    delete (globalThis as any).__CLI_NO_COLOR;
-  });
-
-  function makeReport(
-    overrides: Partial<AuthorStatsReport> = {},
-  ): AuthorStatsReport {
-    return {
-      owner: "anthropics",
-      totalSkills: 10,
-      repos: ["anthropics/skills", "anthropics/more"],
-      categories: { coding: 4, testing: 3, security: 2, devops: 1 },
-      verifiedCount: 7,
-      totalTokens: 25000,
-      topSkills: [
-        { name: "code-review", repo: "anthropics/skills" },
-        { name: "openspec", repo: "anthropics/skills" },
-      ],
-      ...overrides,
-    };
-  }
-
-  it("includes author name and overview", () => {
-    const output = formatAuthorStatsReport(makeReport());
-    expect(output).toContain("Author: anthropics");
-    expect(output).toContain("Total Skills:");
-    expect(output).toContain("10");
-    expect(output).toContain("Repos:");
-    expect(output).toContain("2");
-  });
-
-  it("shows categories with bars", () => {
-    const output = formatAuthorStatsReport(makeReport());
-    expect(output).toContain("Categories");
-    expect(output).toContain("coding");
-    expect(output).toContain("#");
-  });
-
-  it("shows top skills", () => {
-    const output = formatAuthorStatsReport(makeReport());
-    expect(output).toContain("Top Skills");
-    expect(output).toContain("code-review");
-    expect(output).toContain("openspec");
-  });
-
-  it("handles zero skills", () => {
-    const output = formatAuthorStatsReport(
-      makeReport({
-        totalSkills: 0,
-        repos: [],
-        categories: {},
-        verifiedCount: 0,
-        topSkills: [],
-      }),
-    );
-    expect(output).toContain("Author: anthropics");
-    expect(output).toContain("Total Skills:");
-    expect(output).toContain("0");
-  });
-});
-
-// ─── formatIndexStatsReport ──────────────────────────────────────────────────
-
-describe("formatIndexStatsReport", () => {
-  beforeEach(() => {
-    (globalThis as any).__CLI_NO_COLOR = true;
-  });
-  afterEach(() => {
-    delete (globalThis as any).__CLI_NO_COLOR;
-  });
-
-  function makeReport(
-    overrides: Partial<IndexStatsReport> = {},
-  ): IndexStatsReport {
-    return {
-      totalRepos: 5,
-      totalSkills: 50,
-      totalAuthors: 3,
-      categoryDistribution: {
-        coding: 15,
-        testing: 10,
-        security: 8,
-        devops: 7,
-        frontend: 5,
-        general: 5,
-      },
-      verifiedCount: 20,
-      totalTokens: 100000,
-      avgTokensPerSkill: 2000,
-      ...overrides,
-    };
-  }
-
-  it("includes overview stats", () => {
-    const output = formatIndexStatsReport(makeReport());
-    expect(output).toContain("Index Statistics");
-    expect(output).toContain("Repos:");
-    expect(output).toContain("5");
-    expect(output).toContain("Skills:");
-    expect(output).toContain("50");
-    expect(output).toContain("Authors:");
-    expect(output).toContain("3");
-    expect(output).toContain("Avg Tokens/Skill:");
-    // Token figures render via formatTokenCount, never as a bare or byte-sized
-    // number (issues #188, #421).
-    expect(output).toContain("~2k tokens");
-  });
-
-  it("shows category distribution with bars", () => {
-    const output = formatIndexStatsReport(makeReport());
-    expect(output).toContain("Category Distribution");
-    expect(output).toContain("coding");
-    expect(output).toContain("#");
-  });
-
-  it("handles zero stats", () => {
-    const output = formatIndexStatsReport(
-      makeReport({
-        totalRepos: 0,
-        totalSkills: 0,
-        totalAuthors: 0,
-        categoryDistribution: {},
-        verifiedCount: 0,
-        totalTokens: 0,
-        avgTokensPerSkill: 0,
-      }),
-    );
-    expect(output).toContain("Index Statistics");
-    expect(output).toContain("0");
-  });
-});
-
-// ─── Attention budget (issue #421) ──────────────────────────────────────────
-
-describe("median", () => {
-  it("returns 0 for an empty list", () => {
-    expect(median([])).toBe(0);
-  });
-
-  it("returns the middle value for an odd-length list", () => {
-    expect(median([5, 1, 3])).toBe(3);
-  });
-
-  it("averages the two middle values for an even-length list", () => {
-    expect(median([1, 2, 3, 4])).toBe(3);
-  });
-});
-
-describe("computeTokenBudget", () => {
-  it("derives resident cost from the description, not the body", () => {
-    const report = computeTokenBudget([
-      makeSkill({ description: "hello world", tokenCount: 5000 }),
-    ]);
-    expect(report.totalResidentTokens).toBe(3);
-    expect(report.totalBodyTokens).toBe(5000);
-    expect(report.totalResidentTokens).not.toBe(report.totalBodyTokens);
-  });
-
-  it("breaks resident and body totals down per provider", () => {
-    const report = computeTokenBudget([
-      makeSkill({
-        dirName: "a",
-        provider: "claude",
-        providerLabel: "Claude Code",
-        description: "aa bb cc",
-        tokenCount: 100,
-      }),
-      makeSkill({
-        dirName: "b",
-        provider: "codex",
-        providerLabel: "Codex",
-        description: "dd",
-        tokenCount: 40,
-      }),
-    ]);
-    const claude = report.byProvider.find((g) => g.key === "claude");
-    const codex = report.byProvider.find((g) => g.key === "codex");
-    expect(claude).toMatchObject({
-      label: "Claude Code",
-      skills: 1,
-      residentTokens: 5,
-      bodyTokens: 100,
-    });
-    expect(codex).toMatchObject({
-      skills: 1,
-      residentTokens: 1,
-      bodyTokens: 40,
-    });
-    // Sorted by resident cost, heaviest first.
-    expect(report.byProvider[0].key).toBe("claude");
-  });
-
-  it("breaks totals down per scope", () => {
-    const report = computeTokenBudget([
-      makeSkill({ dirName: "a", scope: "global", description: "a b c" }),
-      makeSkill({ dirName: "b", scope: "project", description: "d" }),
-    ]);
-    const keys = report.byScope.map((g) => g.key).sort();
-    expect(keys).toEqual(["global", "project"]);
-    expect(report.byScope.find((g) => g.key === "project")!.skills).toBe(1);
-  });
-
-  it("ranks the heaviest resident descriptions first", () => {
-    const report = computeTokenBudget([
-      makeSkill({ name: "small", dirName: "small", description: "a" }),
-      makeSkill({
-        name: "big",
-        dirName: "big",
-        description: "one two three four five six",
-      }),
-    ]);
-    expect(report.heaviestResident[0].name).toBe("big");
-    expect(report.heaviestResident[0].residentTokens).toBeGreaterThan(
-      report.heaviestResident[1].residentTokens,
-    );
-  });
-
-  it("honours the heaviest-resident limit", () => {
-    const skills = Array.from({ length: 5 }, (_, i) =>
-      makeSkill({ name: `s${i}`, dirName: `s${i}`, description: "a b c" }),
-    );
-    expect(computeTokenBudget(skills, 2).heaviestResident).toHaveLength(2);
-  });
-
-  it("caps heaviestResident at 10 and reports the untruncated total", () => {
-    const skills = Array.from({ length: 11 }, (_, i) =>
-      makeSkill({ name: `s${i}`, dirName: `s${i}`, description: "a b c" }),
-    );
-    const report = computeTokenBudget(skills);
-    expect(report.heaviestResident).toHaveLength(10);
-    expect(report.heaviestResidentTotal).toBe(11);
-    expect(report.heaviestResidentTruncated).toBe(true);
-  });
-
-  it("does not mark an exact-limit list as truncated", () => {
-    const skills = Array.from({ length: 10 }, (_, i) =>
-      makeSkill({ name: `s${i}`, dirName: `s${i}`, description: "a b c" }),
-    );
-    const report = computeTokenBudget(skills);
-    expect(report.heaviestResident).toHaveLength(10);
-    expect(report.heaviestResidentTotal).toBe(10);
-    expect(report.heaviestResidentTruncated).toBe(false);
-  });
-
-  it("reports zeroes for an empty installed set", () => {
-    const report = computeTokenBudget([]);
-    expect(report).toMatchObject({
-      totalSkills: 0,
-      totalResidentTokens: 0,
-      totalBodyTokens: 0,
-      medianResidentTokens: 0,
-      heaviestResidentTotal: 0,
-      heaviestResidentTruncated: false,
-    });
-    expect(report.heaviestResident).toEqual([]);
-  });
-});
-
-describe("formatTokenBudgetReport", () => {
-  beforeEach(() => {
-    (globalThis as any).__CLI_NO_COLOR = true;
-  });
-
-  afterEach(() => {
-    delete (globalThis as any).__CLI_NO_COLOR;
-  });
-
-  it("renders every token figure with the `~` approximation prefix", () => {
-    const output = formatTokenBudgetReport(
-      computeTokenBudget([
-        makeSkill({ description: "one two three", tokenCount: 1234 }),
-      ]),
-    );
-    expect(output).toContain("Attention Budget");
-    expect(output).toContain("~5 tokens");
-    expect(output).toContain("~1.2k tokens");
-    // Token totals must never be rendered as disk sizes.
-    expect(output).not.toMatch(/\d\s?(KB|MB|GB)/);
-  });
-
-  it("right-aligns the heaviest-resident token column", () => {
-    // `~9 tokens` and `~11 tokens` differ in width; left-aligning them makes
-    // the trailing `(provider, scope)` column ragged in a magnitude ranking.
-    const output = formatTokenBudgetReport(
-      computeTokenBudget([
-        makeSkill({
-          name: "big",
-          dirName: "big",
-          description: "one two three four five six",
-        }),
-        makeSkill({
-          name: "small",
-          dirName: "small",
-          description: "one two three four five",
-        }),
-      ]),
-    );
-    expect(output).toContain("big    ~11 tokens");
-    expect(output).toContain("small   ~9 tokens");
-  });
-
-  it("handles an empty installed set", () => {
-    expect(formatTokenBudgetReport(computeTokenBudget([]))).toContain(
-      "  No installed skills.",
-    );
-  });
-
-  it("mentions hidden heaviest rows when the list is truncated", () => {
-    const skills = Array.from({ length: 11 }, (_, i) =>
-      makeSkill({ name: `s${i}`, dirName: `s${i}`, description: "a b c" }),
-    );
-    expect(formatTokenBudgetReport(computeTokenBudget(skills))).toContain(
-      "… 1 more not shown — use --json for the full list",
-    );
   });
 });
